+++ src/general-private.h
@@ -50,7 +50,7 @@
 #endif
 
 #define MAKE_ARGB_RGB(r,g,b)		(((r&0xff) << 16) | ((g&0xff) <<  8) | (b&0xff))
-#define MAKE_ARGB_ARGB(a,r,g,b)		(((a&0xff) << 24) | ((r&0xff) << 16) | ((g&0xff) << 8) | (b&0xff))
+#define MAKE_ARGB_ARGB(a,r,g,b)		(((guint32)(a&0xff) << 24) | ((r&0xff) << 16) | ((g&0xff) << 8) | (b&0xff))
 
 #if !defined(min)
 #define min(a,b)			(a < b ? a : b)
